+++ app_python/requirements.txt
@@ -9,3 +9,7 @@ uvicorn[standard]==0.18.3
 
 # Code style
 black==22.6.0
+
+# Testing
+pytest==7.1.3
+pytest-asyncio==0.19.0
